# feat(query): improve performance of count distinct.